refactor: remove environment-url
Whenever a list of tags were passed to the input, the environment-url
was invalid. Since the environment key is on a job-level, pre-computing
the correct url would require to run a prior job, which is not very
efficient.
Since the url is mainly used for generating a link after deployment,
this is not critical to remove. If a correct URL is required, this
should be done with the actual artifact URL coming from the
docker-composite-action, which can be set during the job by interacting
with the github-api. For now, this is not implemented.